+++ CHANGELOG.md
@@ -3,6 +3,17 @@
 ## Next release
 
 - feat: add integration tests for block production
+- test: tests added for verify and apply task in l2 sync
+- fix: UDC cairo 0 migration & events logic fix
+- fix: response of spec version rpc call fixed
+- tests: integration test for l2 fetch task added
+- refactor: calculate class hashes in devnet
+- feat: add config file and preset configure chain
+- refactor: change default chain id and add custom flag to override
+- fix: generate a fixed set of public and private keys for devnet
+- fix: defaulted l1 gas price in devnet mode
+- fix: fixed anvil port value in tests
+- fix: flaky tests in gas price worker fixed
 - ci: add coveralls report
 - test: added tests for declare and deploy transactions
 - fix: pending block must always be returned in rpc even if none is in db